Working around bug in CoffeeScript Redux compiler
I noticed that in certain cases, the AST produced by the CSR compiler doesn't add this `kind: 'init'` property and would error out in the `default:` case below. I'm assuming that `'init'` is a basic thing you'd see almost all the time, so I figured I'd just add it in.